ci!: hermetic CI on nix across three targets, with published coverage
This repository had no CI. Every check ran on one machine with one compiler --
Apple clang on arm64 -- so "the C is portable" and "the build is reproducible"
were both untested claims. They are now checked on every push.

WHAT RUNS. Five build legs (linux-x64, linux-arm64, macos-arm64; gcc and clang
on Linux, clang on Darwin), the Haskell library and its 2428-test suite, and a
coverage job. Toolchains come from a flake pinned to a nixpkgs COMMIT rather
than a branch, so a fixed source revision resolves a fixed compiler. Nix pins
the compiler only -- Hackage packages still come from cabal, so
cabal.project.freeze with an index-state pin is what makes the Haskell side
reproducible, and it lands here too.

WHAT IT FOUND. Every one of these was a real defect, invisible while a single
compiler was the only reader:

  * [[unsequenced]] / [[reproducible]] were misplaced at all 63 sites. C23
    requires them on the function declarator; every use here is a prefix.
    Clang 19 does not implement the attribute, silently took the GNU fallback,
    and said nothing; GCC 15 does implement it and refused.
  * Thirteen diagnostic suppressions around vendored StringZilla were spelled
    `#pragma clang diagnostic`, which GCC ignores outright -- so on GCC none of
    them applied.
  * POSIX was invisible under strict -std=c23 across 39 files: glibc withholds
    it, and the feature-test macros were absent or too late to matter.
  * An enum-compare between two anonymous enums, truncation at BOTH ends of the
    shrinker's report chain, and a file-scope static unused in one suite.

WHAT IS ASSERTED, NOT DESCRIBED. Three invariants are now checked by the build
rather than written down and hoped for:

  * wok_base.h #errors if it is reached after a system header, because a
    feature-test macro that arrives late is merely decorative. Documenting the
    ordering rule was not enough -- the rule was already documented and already
    violated in fifteen files, and the check found twenty-four more.
  * The Haskell job fails if the test log contains "skipped (WOK_WOKPARSE
    unset)". Three groups are gated on that variable and, unset, they do not
    fail: they register a passing test. 2318 green without it, 2428 with. An
    unguarded run reports success while dropping the C-vs-Haskell differential.
  * coverage.json is checked against lcov's own summary of the same tracefile
    and the build stops on any disagreement. The two silently differed once --
    365 functions against 686 -- because the parser knew only the classic
    FN:/FNDA: spelling and lcov 2.x writes FNL:/FNA:. A percentage that is
    quietly wrong is worse than none, and a tool cannot notice.

COVERAGE is published to coverage.wokml.org: 94.7% of lines, 97.0% of
functions over 21 files. Alongside the HTML sit coverage.json (per-file
totals plus the individual uncovered lines and functions), build.json
(provenance) and the raw tracefile, so a tool reads one document instead of
scraping 86 files of markup. Every artifact is stamped with the commit it
measured and the run that produced it, and the report is dated by its commit
rather than wall-clock so it stays byte-reproducible.

The publish step asserts WHERE it is about to write before writing: an earlier
attempt deployed to the project serving wokml.org and replaced the main site.
Branch guards would not have prevented that -- the step was right about when to
run and wrong about its target -- so the target's domain is now verified
against Netlify's own record, and a mismatch fails the job.

BREAKING CHANGE: the C toolchain floor is now GCC 15 or Clang 19, both
measured rather than assumed. GCC 14 compiles every C23 feature this code uses
but still reports __STDC_VERSION__ as 202000L, which include/wok_base.h
refuses; Clang 18 has no constexpr in C, which wok_token.c requires.

Squashed from 20 commits (6ea77ba..a20e33c) on ci/hermetic-gcc-clang.

+++ grammar/c/wok_print.c
@@ -39,13 +39,17 @@
 // parentheses is deliberately NOT done: punctuation the author did not write
 // is a worse trade than one long line.
 
+// The prelude comes FIRST: it carries the POSIX feature-test macros, which
+// have no effect once a system header has been read. wok_base.h hard-errors
+// if it is reached too late.
+#include "wok_base.h"
+
 #include "wok_print.h"
 
 #include <inttypes.h>
 #include <stdint.h>
 #include <string.h>
 
-#include "wok_base.h"
 #include "wok_token.h"
 #include "wok_trivia.h"
 
@@ -405,8 +409,13 @@ enum { EP_EXPR = 0, EP_CHAIN = 1, EP_NEG = 2, EP_APP = 3, EP_ATOM = 4 };
 // path or an H_ChainOp -- where a bracket is SYNTAX, not grouping -- so the
 // invariant is asserted rather than described.
 enum { PREC_FIXED = 4 };
-static_assert(PREC_FIXED >= TP_ATOM && PREC_FIXED >= PP_ATOM &&
-                  PREC_FIXED >= EP_ATOM,
+// The casts carry weight: PREC_FIXED and the *_ATOM rungs are enumerators of
+// DIFFERENT anonymous enums, and GCC's -Wenum-compare rejects comparing two
+// unrelated enum types. Comparing their integer values is precisely what is
+// meant here, so say so.
+static_assert((int)PREC_FIXED >= (int)TP_ATOM &&
+                  (int)PREC_FIXED >= (int)PP_ATOM &&
+                  (int)PREC_FIXED >= (int)EP_ATOM,
               "PREC_FIXED must top every ladder; see the note above");
 
 typedef struct {
